midjourney-automation/requirements.txt
selenium==4.16.0
discord.py==2.3.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
Pillow==10.1.0
requests==2.31.0 